+++ src/bitcoin/validation.rs
@@ -1,4 +1,4 @@
-//! Bitcoin transaction validation [AIS-3][BPC-3][DAO-4]
+//! Bitcoin transaction validation [AIS-3][BPC-3][DAO-3]
 
 use bitcoin::{Transaction, Block, BlockHeader};
 use thiserror::Error;